+++ client/out/src/extension.js
@@ -25,7 +25,10 @@ async function activate(context) {
     // Options to control the language client
     const clientOptions = {
         // Register the server for plain text documents
-        documentSelector: [{ scheme: 'file', language: 'tinderbox-action-code' }],
+        documentSelector: [
+            { scheme: 'file', language: 'tinderbox-action-code' },
+            { scheme: 'file', language: 'tinderbox-export-code' }
+        ],
         synchronize: {
             // Notify the server about file changes to '.clientrc files contained in the workspace
             fileEvents: vscode_1.workspace.createFileSystemWatcher('**/.clientrc')